Refactor monster repository
